Fix int64_t/long-long type-deduction mismatches for NDK libc++ (P3)
Two std::min/std::max calls failed to compile under the NDK's libc++
(clang++ --target=aarch64-none-linux-android28) with "no matching
function... deduced conflicting types": both templates require an EXACT
type match between arguments, and int64_t/uint64_t are typedef'd to
`long` on this ABI while a bare `4ull` literal or a
std::chrono::duration::rep is `long long` - two distinct 64-bit integer
types, not implicitly convertible for template argument deduction (this
is a real platform ABI difference, not a NDK bug).

dvd.cpp: took PR #37's exact fix (master hadn't diverged since PR #37
branched - 0-diff against the merge-base) - UINT64_C(4) instead of
4ull, plus an explicit uint64_t cast on the other std::max argument.

audio.cpp: PR #37 never hit this one (its libstdc++/glibc combination
apparently deduces these particular types compatibly - a real
difference between standard library implementations, not something to
question), so no upstream fix existed to take; applied the same
explicit-cast pattern directly.

Verified: compiled clean via the real NDK clang for arm64-v8a as part of
the runtime_common unity build (caught these as real build errors, not
theoretical - this is genuinely the first time this file has been
compiled against libc++ rather than libstdc++/MSVC STL).

+++ runtime/src/hle/audio/audio.cpp
@@ -509,7 +509,11 @@ int64_t ConsumeAudioPollDeltaMicros()
     // kMaxBlocksPerTick per pass. Never return early on a zero delta: two scheduler passes
     // can land in the same microsecond and the backlog still needs servicing.
     constexpr int64_t kMaxPollDeltaMicros = 100'000;
-    return std::min(elapsed, kMaxPollDeltaMicros);
+    // std::chrono::microseconds::rep is a distinct type from int64_t on some platforms (e.g.
+    // NDK libc++: `long long` vs int64_t's `long`) even though both are 64-bit - std::min<T>
+    // requires an exact type match between its two arguments, not just equal width, so this
+    // needs an explicit cast rather than relying on implicit conversion.
+    return std::min(static_cast<int64_t>(elapsed), kMaxPollDeltaMicros);
 }
 
 } // namespace
